Present memory sizes and DNA value in system info
Present the following memory dimensions in the system info:
- Main memory (SRAM).
- HyperRAM.
- Portion of the HyperRAM that is supported by tag bits an
  may thus store capabilities.

Make the DNA value of the FPGA available in the system info
IP block; this is a 57-bit value programmed into the FPGA
which is described as 'most often unique', although there may
be up to 32 devices in the FPGA family which have the same
DNA value.

Model the DNA_PORT/value in simulation, using a single value
for all simulations.